Use Diesel to persist data in a SQLite database

+++ src/models.rs
@@ -1,7 +1,12 @@
+use diesel::Insertable;
+use diesel::Queryable;
+use crate::schema::todo_items;
+
 // Represents a to-do list item with its details
-#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
+#[derive(Clone, Debug, serde::Serialize, serde::Deserialize, Insertable, Queryable)]
+#[diesel(table_name = todo_items)]
 pub struct TodoItem {
-    pub id: u64,
+    pub id: i32,
     pub title: String,
     pub description: String,
     pub completed: bool,
